FindingMiniumDiffElemForGivenKeyInSortedArray
This sample describes how to get the minimum different element in sorted array for a given key
We need to do a binary search. and  find the key in the array.
If Key is found that is the minimum diff element
If Key is not given we try to do binary search till the start and end indexes cross each other

When the while loop exits, we have the start and end indexes pointing to the  right and left neighbour for the gievn key.
In other words start and end indexes are pointing to ceil and floor value of the given key.
FInd the absolute diff for the key with the floor and ceil element.  whichever is he minimum that is the index of the element with the minimum difference from the key.